Factorize a bit
