+++ Extensions/TileMap/simpletilemapruntimeobject.ts
@@ -38,14 +38,9 @@ namespace gdjs {
    * @category Objects > Tile Map
    */
   export class SimpleTileMapRuntimeObject
-    extends gdjs.RuntimeObject
+    extends gdjs.AbstractTileMapRuntimeObject
     implements gdjs.TileMap, gdjs.Resizable, gdjs.Scalable, gdjs.OpacityHandler
   {
-    /**
-     * A reusable Point to avoid allocations.
-     */
-    private static readonly workingPoint: FloatPoint = [0, 0];
-
     _opacity: float = 255;
     _atlasImage: string;
     _tileMapManager: gdjs.TileMap.TileMapRuntimeManager;
@@ -59,13 +54,8 @@ namespace gdjs {
     _initialTileMapAsJsObject: TileMapHelper.EditableTileMapAsJsObject;
     readonly _initialTilesWithHitBox: number[];
     _isTileMapDirty: boolean = false;
-    _sceneToTileMapTransformation: gdjs.AffineTransformation =
-      new gdjs.AffineTransformation();
-    _tileMapToSceneTransformation: gdjs.AffineTransformation =
-      new gdjs.AffineTransformation();
     _collisionTileMap: gdjs.TileMap.TransformedCollisionTileMap | null = null;
     _hitBoxTag: string = 'collision';
-    private _transformationIsUpToDate: boolean = false;
 
     // TODO: Add a debug mode like for TileMapCollisionMaskRuntimeObject to draw?
 
@@ -269,7 +259,7 @@ namespace gdjs {
               );
           }
 
-          this._transformationIsUpToDate = false;
+          this.invalidateTransformation();
           this.updateTransformation();
           this.invalidateHitboxes();
         }
@@ -355,7 +345,7 @@ namespace gdjs {
     setWidth(width: float): void {
       if (this.getWidth() === width) return;
 
-      this._transformationIsUpToDate = false;
+      this.invalidateTransformation();
       this._renderer.setWidth(width);
       this.invalidateHitboxes();
     }
@@ -363,7 +353,7 @@ namespace gdjs {
     setHeight(height: float): void {
       if (this.getHeight() === height) return;
 
-      this._transformationIsUpToDate = false;
+      this.invalidateTransformation();
       this._renderer.setHeight(height);
       this.invalidateHitboxes();
     }
@@ -371,7 +361,6 @@ namespace gdjs {
     setSize(newWidth: float, newHeight: float): void {
       this.setWidth(newWidth);
       this.setHeight(newHeight);
-      this._transformationIsUpToDate = false;
     }
 
     /**
@@ -393,7 +382,6 @@ namespace gdjs {
     setScale(scale: float): void {
       this.setScaleX(scale);
       this.setScaleY(scale);
-      this._transformationIsUpToDate = false;
     }
 
     /**
@@ -409,7 +397,7 @@ namespace gdjs {
 
       this._renderer.setScaleX(scaleX);
       this.invalidateHitboxes();
-      this._transformationIsUpToDate = false;
+      this.invalidateTransformation();
     }
 
     /**
@@ -425,25 +413,25 @@ namespace gdjs {
 
       this._renderer.setScaleY(scaleY);
       this.invalidateHitboxes();
-      this._transformationIsUpToDate = false;
+      this.invalidateTransformation();
     }
 
     setX(x: float): void {
       super.setX(x);
       this._renderer.updatePosition();
-      this._transformationIsUpToDate = false;
+      this.invalidateTransformation();
     }
 
     setY(y: float): void {
       super.setY(y);
       this._renderer.updatePosition();
-      this._transformationIsUpToDate = false;
+      this.invalidateTransformation();
     }
 
     setAngle(angle: float): void {
       super.setAngle(angle);
       this._renderer.updateAngle();
-      this._transformationIsUpToDate = false;
+      this.invalidateTransformation();
     }
 
     setOpacity(opacity: float): void {
@@ -606,276 +594,6 @@ namespace gdjs {
       return true;
     }
 
-    updateTransformation() {
-      if (this._transformationIsUpToDate) {
-        return;
-      }
-
-      this._tileMapToSceneTransformation.setToIdentity();
-      this._tileMapToSceneTransformation.translate(this.getX(), this.getY());
-      this._tileMapToSceneTransformation.rotateAround(
-        (this.getAngle() * Math.PI) / 180,
-        this.getCenterX(),
-        this.getCenterY()
-      );
-      this._tileMapToSceneTransformation.scale(
-        Math.abs(this._renderer.getScaleX()),
-        Math.abs(this._renderer.getScaleY())
-      );
-
-      if (this._collisionTileMap) {
-        const collisionTileMapTransformation =
-          this._collisionTileMap.getTransformation();
-        collisionTileMapTransformation.copyFrom(
-          this._tileMapToSceneTransformation
-        );
-        this._collisionTileMap.setTransformation(
-          collisionTileMapTransformation
-        );
-      }
-      this._sceneToTileMapTransformation.copyFrom(
-        this._tileMapToSceneTransformation
-      );
-      this._sceneToTileMapTransformation.invert();
-      this._transformationIsUpToDate = true;
-    }
-
-    getSceneXCoordinateOfTileCenter(
-      columnIndex: integer,
-      rowIndex: integer
-    ): float {
-      const sceneCoordinates: FloatPoint =
-        SimpleTileMapRuntimeObject.workingPoint;
-      this._tileMapToSceneTransformation.transform(
-        [
-          (columnIndex + 0.5) * this._tileSize,
-          (rowIndex + 0.5) * this._tileSize,
-        ],
-        sceneCoordinates
-      );
-      return sceneCoordinates[0];
-    }
-
-    getSceneYCoordinateOfTileCenter(
-      columnIndex: integer,
-      rowIndex: integer
-    ): float {
-      const sceneCoordinates: FloatPoint =
-        SimpleTileMapRuntimeObject.workingPoint;
-      this._tileMapToSceneTransformation.transform(
-        [
-          (columnIndex + 0.5) * this._tileSize,
-          (rowIndex + 0.5) * this._tileSize,
-        ],
-        sceneCoordinates
-      );
-      return sceneCoordinates[1];
-    }
-
-    /**
-     * The returned array is alway the same.
-     */
-    getGridCoordinatesFromSceneCoordinates(
-      x: float,
-      y: float
-    ): [integer, integer] {
-      this.updateTransformation();
-
-      const result = SimpleTileMapRuntimeObject.workingPoint;
-      this._sceneToTileMapTransformation.transform([x, y], result);
-
-      result[0] = Math.floor(result[0] / this._tileSize);
-      result[1] = Math.floor(result[1] / this._tileSize);
-      return result;
-    }
-
-    getColumnIndexAtPosition(x: float, y: float): integer {
-      return this.getGridCoordinatesFromSceneCoordinates(x, y)[0];
-    }
-
-    getRowIndexAtPosition(x: float, y: float): integer {
-      return this.getGridCoordinatesFromSceneCoordinates(x, y)[1];
-    }
-
-    getTileAtPosition(x: float, y: float): integer {
-      const [columnIndex, rowIndex] =
-        this.getGridCoordinatesFromSceneCoordinates(x, y);
-      return this.getTileAtGridCoordinates(columnIndex, rowIndex);
-    }
-
-    getTileAtGridCoordinates(columnIndex: integer, rowIndex: integer): integer {
-      return this.getTileId(columnIndex, rowIndex, 0);
-    }
-
-    setTileAtPosition(tileId: number, x: float, y: float) {
-      const [columnIndex, rowIndex] =
-        this.getGridCoordinatesFromSceneCoordinates(x, y);
-      this.setTileAtGridCoordinates(tileId, columnIndex, rowIndex);
-    }
-
-    setTileAtGridCoordinates(
-      tileId: number,
-      columnIndex: integer,
-      rowIndex: integer
-    ) {
-      if (!this._tileMap) {
-        return;
-      }
-      const layer = this._tileMap.getTileLayer(this._layerIndex);
-      if (!layer) {
-        return;
-      }
-      const oldTileId = layer.getTileId(columnIndex, rowIndex);
-      if (tileId === oldTileId) {
-        return;
-      }
-      layer.setTile(columnIndex, rowIndex, tileId);
-
-      if (this._collisionTileMap) {
-        const oldTileDefinition =
-          oldTileId !== undefined && this._tileMap.getTileDefinition(oldTileId);
-        const newTileDefinition = this._tileMap.getTileDefinition(tileId);
-        const hadFullHitBox =
-          !!oldTileDefinition &&
-          oldTileDefinition.hasFullHitBox(this._hitBoxTag);
-        const haveFullHitBox =
-          !!newTileDefinition &&
-          newTileDefinition.hasFullHitBox(this._hitBoxTag);
-        if (hadFullHitBox !== haveFullHitBox) {
-          this._collisionTileMap.invalidateTile(
-            this._layerIndex,
-            columnIndex,
-            rowIndex
-          );
-        }
-      }
-      this._isTileMapDirty = true;
-    }
-
-    flipTileOnYAtPosition(x: float, y: float, flip: boolean) {
-      const [columnIndex, rowIndex] =
-        this.getGridCoordinatesFromSceneCoordinates(x, y);
-      this.flipTileOnYAtGridCoordinates(columnIndex, rowIndex, flip);
-    }
-
-    flipTileOnXAtPosition(x: float, y: float, flip: boolean) {
-      const [columnIndex, rowIndex] =
-        this.getGridCoordinatesFromSceneCoordinates(x, y);
-      this.flipTileOnXAtGridCoordinates(columnIndex, rowIndex, flip);
-    }
-
-    flipTileOnYAtGridCoordinates(
-      columnIndex: integer,
-      rowIndex: integer,
-      flip: boolean
-    ) {
-      this.flipTileOnY(columnIndex, rowIndex, 0, flip);
-      this._isTileMapDirty = true;
-      // No need to invalidate hit boxes since at the moment, collision mask
-      // cannot be configured on each tile.
-    }
-
-    flipTileOnXAtGridCoordinates(
-      columnIndex: integer,
-      rowIndex: integer,
-      flip: boolean
-    ) {
-      this.flipTileOnX(columnIndex, rowIndex, 0, flip);
-      this._isTileMapDirty = true;
-      // No need to invalidate hit boxes since at the moment, collision mask
-      // cannot be configured on each tile.
-    }
-
-    isTileFlippedOnXAtPosition(x: float, y: float) {
-      const [columnIndex, rowIndex] =
-        this.getGridCoordinatesFromSceneCoordinates(x, y);
-
-      return this.isTileFlippedOnX(columnIndex, rowIndex, 0);
-    }
-
-    isTileFlippedOnXAtGridCoordinates(columnIndex: integer, rowIndex: integer) {
-      return this.isTileFlippedOnX(columnIndex, rowIndex, 0);
-    }
-
-    isTileFlippedOnYAtPosition(x: float, y: float) {
-      const [columnIndex, rowIndex] =
-        this.getGridCoordinatesFromSceneCoordinates(x, y);
-
-      return this.isTileFlippedOnY(columnIndex, rowIndex, 0);
-    }
-
-    isTileFlippedOnYAtGridCoordinates(columnIndex: integer, rowIndex: integer) {
-      return this.isTileFlippedOnY(columnIndex, rowIndex, 0);
-    }
-
-    removeTileAtPosition(x: float, y: float) {
-      const [columnIndex, rowIndex] =
-        this.getGridCoordinatesFromSceneCoordinates(x, y);
-      this.removeTileAtGridCoordinates(columnIndex, rowIndex);
-    }
-
-    removeTileAtGridCoordinates(columnIndex: integer, rowIndex: integer) {
-      if (!this._tileMap) {
-        return;
-      }
-      const layer = this._tileMap.getTileLayer(this._layerIndex);
-      if (!layer) {
-        return;
-      }
-      const oldTileId = layer.getTileId(columnIndex, rowIndex);
-      if (oldTileId === undefined) {
-        return;
-      }
-      layer.removeTile(columnIndex, rowIndex);
-      if (this._collisionTileMap) {
-        const oldTileDefinition =
-          oldTileId !== undefined && this._tileMap.getTileDefinition(oldTileId);
-        const hadFullHitBox =
-          !!oldTileDefinition &&
-          oldTileDefinition.hasFullHitBox(this._hitBoxTag);
-        if (hadFullHitBox) {
-          this._collisionTileMap.invalidateTile(
-            this._layerIndex,
-            columnIndex,
-            rowIndex
-          );
-        }
-      }
-      this._isTileMapDirty = true;
-    }
-
-    setGridRowCount(targetRowCount: integer) {
-      if (targetRowCount <= 0) return;
-      if (!this._tileMap) return;
-      this._tileMap.setDimensionY(targetRowCount);
-      if (this._collisionTileMap) {
-        this._collisionTileMap.updateDimensions();
-      }
-      this._isTileMapDirty = true;
-      this.invalidateHitboxes();
-    }
-
-    setGridColumnCount(targetColumnCount: integer) {
-      if (targetColumnCount <= 0) return;
-      if (!this._tileMap) return;
-      this._tileMap.setDimensionX(targetColumnCount);
-      if (this._collisionTileMap) {
-        this._collisionTileMap.updateDimensions();
-      }
-      this._isTileMapDirty = true;
-      this.invalidateHitboxes();
-    }
-
-    getGridRowCount(): integer {
-      if (!this._tileMap) return 0;
-      return this._tileMap.getDimensionY();
-    }
-
-    getGridColumnCount(): integer {
-      if (!this._tileMap) return 0;
-      return this._tileMap.getDimensionX();
-    }
-
     getTilesetColumnCount(): integer {
       return this._columnCount;
     }
@@ -888,57 +606,24 @@ namespace gdjs {
       return this._tileMap;
     }
 
-    /**
-     * @param x The layer column.
-     * @param y The layer row.
-     * @param layerIndex The layer index.
-     * @returns The tile's id.
-     */
-    getTileId(x: integer, y: integer, layerIndex: integer): integer {
-      if (!this._tileMap) return -1;
-      return this._tileMap.getTileId(x, y, layerIndex);
+    getTileMapForEdition(): TileMapHelper.EditableTileMap | null {
+      return this._tileMap;
     }
 
-    /**
-     * @param x The layer column.
-     * @param y The layer row.
-     * @param layerIndex The layer index.
-     * @param flip true if the tile should be flipped.
-     */
-    flipTileOnY(x: integer, y: integer, layerIndex: integer, flip: boolean) {
-      if (!this._tileMap) return;
-      this._tileMap.flipTileOnY(x, y, layerIndex, flip);
+    getCollisionTileMap(): gdjs.TileMap.TransformedCollisionTileMap | null {
+      return this._collisionTileMap;
     }
 
-    /**
-     * @param x The layer column.
-     * @param y The layer row.
-     * @param layerIndex The layer index.
-     * @param flip true if the tile should be flipped.
-     */
-    flipTileOnX(x: integer, y: integer, layerIndex: integer, flip: boolean) {
-      if (!this._tileMap) return;
-      this._tileMap.flipTileOnX(x, y, layerIndex, flip);
+    getLayerIndex(): integer {
+      return this._layerIndex;
     }
 
-    /**
-     * @param x The layer column.
-     * @param y The layer row.
-     * @param layerIndex The layer index.
-     */
-    isTileFlippedOnX(x: integer, y: integer, layerIndex: integer): boolean {
-      if (!this._tileMap) return false;
-      return this._tileMap.isTileFlippedOnX(x, y, layerIndex);
+    getCollisionMaskTag(): string {
+      return this._hitBoxTag;
     }
 
-    /**
-     * @param x The layer column.
-     * @param y The layer row.
-     * @param layerIndex The layer index.
-     */
-    isTileFlippedOnY(x: integer, y: integer, layerIndex: integer): boolean {
-      if (!this._tileMap) return false;
-      return this._tileMap.isTileFlippedOnY(x, y, layerIndex);
+    invalidateTileMap(): void {
+      this._isTileMapDirty = true;
     }
   }
   gdjs.registerObject(
